Use uint64_t for RustBuffer fields to fix 32-bit ARM crash (#344)
Originally by  @sfourdrinier, separated out from #313.

+++ cpp/includes/RustBuffer.h
@@ -11,7 +11,7 @@
 #include <jsi/jsi.h>
 
 struct RustBuffer {
-  size_t capacity;
-  size_t len;
+  uint64_t capacity;
+  uint64_t len;
   uint8_t *data;
 };
